Implementation of get post hooks complete.
